Revert "cast final result to smallint"
This reverts commit f80b62f5b255e5644e1a9d26af97e6ff9a12db54.

+++ priv/repo/migrations/20250613133700_add_and_calculate_elevation_changes.exs
@@ -14,8 +14,8 @@ defmodule TeslaMate.Repo.Migrations.AddAndCalculateElevationChanges do
     WITH elevation_changes AS (
       SELECT
         drive_id,
-        COALESCE(NULLIF(LEAST(SUM(CASE WHEN elevation_diff > 0 THEN elevation_diff ELSE 0 END), 32768), 32768), 0)::SMALLINT as ascent,
-        COALESCE(NULLIF(LEAST(SUM(CASE WHEN elevation_diff < 0 THEN ABS(elevation_diff) ELSE 0 END), 32768), 32768), 0)::SMALLINT as descent
+        COALESCE(NULLIF(LEAST(SUM(CASE WHEN elevation_diff > 0 THEN elevation_diff ELSE 0 END), 32768), 32768), 0) as ascent,
+        COALESCE(NULLIF(LEAST(SUM(CASE WHEN elevation_diff < 0 THEN ABS(elevation_diff) ELSE 0 END), 32768), 32768), 0) as descent
       FROM (
         SELECT
           drive_id,
